requirements.txt
absl-py @ file:///C:/ci/absl-py_1623867344543/work
aiohttp @ file:///C:/ci/aiohttp_1614361212023/work
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall==0.2.0
blinker==1.4
brotlipy==0.7.0
cached-property @ file:///tmp/build/80754af9/cached-property_1600785575025/work
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831756778/work
chardet @ file:///C:/ci/chardet_1605296577955/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
colorama==0.4.4
coverage @ file:///C:/ci/coverage_1614615100481/work
cryptography @ file:///C:/ci/cryptography_1616769504165/work
Cython @ file:///C:/ci/cython_1618435342791/work
decorator==5.0.7
dlib @ file:///D:/bld/dlib_1617441752500/work
flatbuffers @ file:///tmp/build/80754af9/python-flatbuffers_1614345733764/work
gast @ file:///tmp/build/80754af9/gast_1597433534803/work
google-auth @ file:///tmp/build/80754af9/google-auth_1625743431937/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1614884204321/work
h5py @ file:///C:/ci/h5py_1622088609188/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877477759/work
imutils==0.5.4
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
Markdown @ file:///C:/ci/markdown_1614364082838/work
matplotlib-inline==0.1.2
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1607349747897/work
numpy @ file:///D:/bld/numpy_1624400134018/work
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
olefile==0.46
opencv-python==4.5.3.56
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
parso==0.8.2
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1625663294916/work
prompt-toolkit==3.0.18
protobuf==3.14.0
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.9.0
PyJWT @ file:///C:/ci/pyjwt_1619682721924/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy @ file:///C:/ci/scipy_1618856124610/work
six==1.15.0
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_baze37/tensorboard_1622025863084/work/tensorboard-2.5.0-py3-none-any.whl
tensorboard-plugin-wit==1.6.0
tensorflow==2.5.0
tensorflow-estimator @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_baze37/tensorflow-estimator_1622026529081/work/tensorflow_estimator-2.5.0-py2.py3-none-any.whl
termcolor==1.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
wcwidth==0.2.5
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt @ file:///C:/ci/wrapt_1607574570428/work
yarl @ file:///C:/ci/yarl_1606940155993/work
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
